[linux] Change the ib700 watchdog driver to work in kernel space
The general idea here is that I don't want to rely on a userspace
daemon in order to make the watchdog timer work because sometimes we
won't have more than a single process in userspace.

This instead uses a kernel timer to trigger the watchdog.  The trouble
is that it appears that on my machine, kernel timers keep firing after
a panic (and will therefor keep refreshing the watchdog) so I need to
explicitly check for panics and use that signal to stap updating the
watchdog.  This seems super fragile, but I guess at least it's
something...

I think the best approach is probably to use a pair of watchdogs: one
that's enabled when we make it to userspace, and one that's always
enabled.  That way we'll still detect panics before /init runs, but
we'll also be able to reliably detect all panics all the time.
Luckily, there's two watchdogs in QEMU :).