[QEMU] Start QEMU's ib700 watchdog at power on
The idea here is that I want to be able to catch panics that happen
before Linux's watchdog driver even gets initialized.  This just
starts up QEMU with the watchdog enabled and gives Linux 30 seconds to
initialize drivers -- this shouldn't be a problem because we have the
ib700 driver in kernel, but it could be nasty if it's in a module on
disk.