[httpd/test_palloc.c] Improve comments for existing tests
This change is based on feedback on the wiki.

*Much longer comments before each test explaining what each
test does, walking through the sections, and explaining what
the sections are looking for and why they are tested like that.

*I can rewrite the tests to split them up into subtests, but
as far as I can see (perhaps I am wrong), make check will go
through the entire test method before returning unlike JUnit,
and having them the way they are allowes chained/integrated tests.

*Also, most to be replaced areas are now explained in comments.

*Some minor changes to tests for clarity and/or forgotten assertions
making some.

*"actual check of entry" means that instead of checking that a
pointer is null, I figured out how to check the actual string
content in the array and compare it to what is expected, so I am
checking the actual entry of the array.

*I'm currently working off of 30+ hours of no sleep so apologies
for any unclear wordings.

*The only test I don't written that is on the feedback is use same allocation
context from multiple threads. Yes, there should be a test written for this,
will be committed in a separate commit.