[super.c] Commented out division using nr_segment
This is likely an issue of unclear intention of the original coder. The
variable nr_segment is not defined at this point. The original code
seemed to either be commenting out this part (and forgetting the
semi-colon before), or dividing it (with an extraneous slash and a
midding semi-colon after). From this error, it seems to be the former.